CPU Boost 1) Add cpu-boost driver When certain bursty and important events take place, it might take a while for the current cpufreq governor to notice the new load and react to it. That would result in poor user experience. To alleviate this, the cpu-boost driver boosts the frequency of a CPU for a short duration to maintain good user experience while the governor catches up.
Specifically, this commit deals with ensuring that when "important" tasks
migrate from a fast CPU to a slow CPU, the frequency of the slow CPU is
boosted to be at least as high as the fast CPU for a short duration.

Since this driver enforces the boost by hooking into standard cpufreq
ADJUST notifiers, it has several advantages:
- More portable across kernel versions where the cpufreq internals might
  have been rewritten.
- Governor agnostic and hence works with multiple governors like
  conservative, ondemand, interactive, etc.
- Does not affect the sampling period/logic of existing governors.
- Can have the boost period adjusted independent of governor sampling
  period.

2)Add a sync limit to cpu-boost

Perform frequency synchronization only when source CPU's frequency
is less than sync_threshold, else sync to the sync_threshold.

3) cpufreq: update cpu-boost driver to make it behave with
 cpufreq hardlimit (Yank555.lu)